Combine items for location array item
